+++ Geode/IR/IInstructionArg.cs
@@ -6,7 +6,7 @@ namespace Geode.IR
 		/// Empty string if no name
 		/// </summary>
 		public string Name { get; }
-		public HashSet<ValueRef> Dependencies { get; }
+		public IReadOnlySet<ValueRef> Dependencies { get; }
 		public void ReplaceValue(ValueRef value, ValueRef with);
 	}
 }