+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Removed ARM Linux support due to packaging issues.
+* Fixed issue with constant comparison operators always returning false.